+++ GroupProject/app/src/main/java/group1/pittapi/SportsReceiver.java
@@ -21,32 +21,37 @@ public class SportsReceiver extends BroadcastReceiver {
             info = intent.getExtras().getString("info");
             final SportEvent se = fromJsonString(info);
 
-            // Show Dialog with prompt to open this app.
-            AlertDialog.Builder builder = new AlertDialog.Builder(context);
-            builder.setTitle("Sports Update")
-                    .setMessage(String.format(Locale.US, "New update to %s game against %s.", se.sport, se.opponent_name));
-            builder.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int which) {
-                    dialog.dismiss();
-                }
-            }).setPositiveButton("See More", new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int which) {
-                    Intent intent1 = new Intent(context, PittScores.class);
-                    intent1.putExtra("OppName", se.opponent_name);
-                    intent1.putExtra("OppScore", se.opponent_score);
-                    intent1.putExtra("PittScore", se.score);
-                    context.startActivity(intent1);
-                }
-            });
-            AlertDialog dialog = builder.create();
-            dialog.show();
+
+            Intent intent1 = new Intent(context, PittScores.class);
+            intent1.putExtra("OppName", se.opponent_name);
+            intent1.putExtra("OppScore", se.opponent_score);
+            intent1.putExtra("PittScore", se.score);
+            context.startActivity(intent1);
+
+//            // Show Dialog with prompt to open this app.
+//            AlertDialog.Builder builder = new AlertDialog.Builder(context);
+//            builder.setTitle("Sports Update")
+//                    .setMessage(String.format(Locale.US, "New update to %s game against %s.", se.sport, se.opponent_name));
+//            builder.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
+//                @Override
+//                public void onClick(DialogInterface dialog, int which) {
+//                    dialog.dismiss();
+//                }
+//            }).setPositiveButton("See More", new DialogInterface.OnClickListener() {
+//                @Override
+//                public void onClick(DialogInterface dialog, int which) {
+//                    Intent intent1 = new Intent(context, PittScores.class);
+//                    intent1.putExtra("OppName", se.opponent_name);
+//                    intent1.putExtra("OppScore", se.opponent_score);
+//                    intent1.putExtra("PittScore", se.score);
+//                    context.startActivity(intent1);
+//                }
+//            });
+//            AlertDialog dialog = builder.create();
+//            dialog.show();
         } catch (Exception e) {
             e.printStackTrace();
         }
-
-
     }
 
     private class SportEvent {
